+++ src/main/server/controllers/ped.ts
@@ -4,6 +4,7 @@ import { PedOptions } from '@Shared/types/pedOptions.js';
 import { Events } from '../../shared/events/index.js';
 import * as Utility from '@Shared/utility/index.js';
 import { useRebar } from '../index.js';
+import { OmitFirstArg } from '../../shared/types/omitFirstArg.js';
 
 const sessionKey = 'ped:uid';
 
@@ -48,7 +49,6 @@ type PedNatives = Pick<
     | 'taskWanderStandard'
 >;
 
-type OmitFirstArg<F> = F extends (x: any, ...args: infer P) => infer R ? (...args: P) => R : never;
 type PedDeathCallback = (uid: string, killer: alt.Entity, weaponHash: number) => void;
 
 const peds: Map<string, ReturnType<typeof usePed>> = new Map();