+++ src/main/server/document/vehicle.ts
@@ -1,9 +1,14 @@
 import * as alt from 'alt-server';
-import { Vehicle } from '@Shared/types/vehicle.js';
-import { KnownKeys } from '@Shared/utilityTypes/index.js';
+import { Vehicle as VehicleDocument } from '@Shared/types/vehicle.js';
 import { useDatabase } from '@Server/database/index.js';
 import { CollectionNames, KeyChangeCallback } from './shared.js';
 import { useRebar } from '../index.js';
+
+declare module 'alt-server' {
+    export interface ICustomEmitEvent {
+        vehicleBound: (vehicle: alt.Vehicle, document: VehicleDocument) => void;
+    }
+}
 
 const Rebar = useRebar();
 const sessionKey = 'document:vehicle';
@@ -30,14 +35,14 @@ export function useVehicle(vehicle: alt.Vehicle) {
      * Return current vehicle data and their associated Vehicle object.
      *
      * @template T
-     * @return {(T & Vehicle) | undefined}
+     * @return {VehicleDocument | undefined}
      */
-    function get<T = {}>(): (T & Vehicle) | undefined {
+    function get(): VehicleDocument | undefined {
         if (!vehicle.hasMeta(sessionKey)) {
             return undefined;
         }
 
-        return <T & Vehicle>vehicle.getMeta(sessionKey);
+        return <VehicleDocument>vehicle.getMeta(sessionKey);
     }
 
     /**
@@ -48,7 +53,7 @@ export function useVehicle(vehicle: alt.Vehicle) {
      * @param {(keyof KnownKeys<Vehicle & T>)} fieldName
      * @return {ReturnType | undefined}
      */
-    function getField<T = {}, ReturnType = any>(fieldName: keyof KnownKeys<Vehicle & T>): ReturnType | undefined {
+    function getField<K extends keyof VehicleDocument>(fieldName: K): VehicleDocument[K] | undefined {
         if (!vehicle.hasMeta(sessionKey)) {
             return undefined;
         }
@@ -65,13 +70,13 @@ export function useVehicle(vehicle: alt.Vehicle) {
      * @param {*} value
      * @return {void}
      */
-    async function set<T = {}, Keys = keyof KnownKeys<Vehicle & T>>(fieldName: Keys, value: any) {
+    async function set<K extends keyof VehicleDocument>(fieldName: K, value: VehicleDocument[K]) {
         if (!vehicle.hasMeta(sessionKey)) {
             return undefined;
         }
 
         const typeSafeFieldName = String(fieldName);
-        let data = vehicle.getMeta(sessionKey) as T & Vehicle;
+        let data = vehicle.getMeta(sessionKey) as VehicleDocument;
         let oldValue = undefined;
 
         if (data[typeSafeFieldName]) {
@@ -101,12 +106,12 @@ export function useVehicle(vehicle: alt.Vehicle) {
      * @param {(Partial<Vehicle & T>)} fields
      * @returns {void}
      */
-    async function setBulk<T = {}, Keys = Partial<Vehicle & T>>(fields: Keys) {
+    async function setBulk(fields: Partial<VehicleDocument>) {
         if (!vehicle.hasMeta(sessionKey)) {
             return undefined;
         }
 
-        let data = vehicle.getMeta(sessionKey) as Vehicle & T;
+        let data = vehicle.getMeta(sessionKey) as VehicleDocument;
 
         const oldValues = {};
 
@@ -156,13 +161,13 @@ export function useVehicleBinder(vehicle: alt.Vehicle) {
      *
      * @param {Vehicle & T} document
      */
-    function bind<T = {}>(document: Vehicle & T, syncVehicle = true): ReturnType<typeof useVehicle> | undefined {
+    function bind(document: VehicleDocument, syncVehicle = true): ReturnType<typeof useVehicle> | undefined {
         if (!vehicle.valid) {
             return undefined;
         }
 
         vehicle.setMeta(sessionKey, document);
-        Rebar.events.useEvents().invoke('vehicle-bound', vehicle, document);
+        alt.emit('vehicleBound', vehicle, document);
 
         if (syncVehicle) {
             Rebar.vehicle.useVehicle(vehicle).sync();
@@ -202,7 +207,7 @@ export function useVehicleEvents() {
      * @param {KeyChangeCallback<alt.Vehicle>} callback
      * @return {void}
      */
-    function on<T = {}>(fieldName: keyof KnownKeys<Vehicle & T>, callback: KeyChangeCallback<alt.Vehicle>) {
+    function on<K extends keyof VehicleDocument>(fieldName: K, callback: KeyChangeCallback<alt.Vehicle>) {
         const actualFieldName = String(fieldName);
 
         if (typeof callbacks[actualFieldName] === 'undefined') {